DN_ServerApp/pyproject.toml
[tool.poetry]
name = "dn-serverapp"
version = "0.1.0"
description = ""
authors = ["izalitzilant <izalitzilant@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.114.0"}
aiofiles = "^24.1.0"
pymongo = "^4.8.0"
hurry-filesize = "^0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
